Move ITransportHandler to transport.py and rename it as BaseTransportHandler
